US128094 Use HMC for View All Work screen
This is a conversion of the `d2l-all-assessments` component to instead
use the non-collapsed version of the HMC `d2l-w2d-collections`
component, plus a date selector that changes the date of which
activities are currently showing the W2D component.

This required some ancillary changes here and there, but was a fairly
straightforward replacement. Once we are no longer using the
`d2l-upcoming-assessments` component (which has already been replaced by
the collapsed W2D component in BfP), we can likely clean out a bunch of
stuff from this repo, as it will effectively just be this
`d2l-all-assessments` component - could even consider moving it back
into the `parent-portal-ui` repo.