Update dependencies
Was seeing some ESLint failures in #179, and realized that we're pretty
out of date. Sure enough, updating that cleared the errors, so figured
might as well update everything else while I'm here. This is mostly
devDependency changes, with the exceptions being a minor bump to
`siren-parser`, and removing `whatwg-fetch`, which was unused.