US128094 Fix date handling
There were two issues here causing weird behaviour with the date
selector. First, by setting the `value` on the selector we create some
strange cyclical behaviour, exacerbated by the problem outlined below.
To avoid this, but to also avoid a blank date selector initially, we
just set the date to "today" via a throwaway value. Once a date has been
selected by a user, at that point it will just display the selected
date.

The second issue was mostly apparently when selecting a date at the
beginning of end of a week (i.e. Sunday or Saturday), as the offsets
along with the cyclical behaviour mentioned above meant the calendar
would go back a week every time you opened it. This was solved by adding
the offset to the date value the `d2l-date-input` gives us, then working
with that. This has the effect of turning the midnight-UTC the selector
gives us into an equivalent datetime that is midnight-UTC but in the
local timezone.